10 July 2025/Research_Agent/requirements.txt
crewai
streamlit
python-dotenv
langchain
langchain-community
